remove limit on number of files updated at once
Since calculating the hash tables is the costliest part of updating a
directory it makes little sense to limit the number of files being
processed at once. That only means the hash tables have to be created
more often and updates take much more time in total. Since it's all
done in an asynchronous worker thread anyway it makes more sense to
process as much as possible in one go, even if it takes a second
longer.