remove dependencies on rust nightly
This commit comprises changes made by @bhavidsharma in PR #91 without
the other formatting changes.
https://github.com/rabite0/hunter/pull/91

It took quite a while to scroll through and pick out the important bits.

This commit wasn't written with Bhavit but I thought it important to
include them as a co-author as it was originally their changes.

Co-authored-by: Bhavit Sharma <bhavitsharma@google.com>